Catch unquoted strings
Twig will (rightfully) go nuts if you ask an attribute it doesn't know
about. Since we seem to assume that either data or value exists, we
don't catch the case where we have neither. Example:

  {% trans foobar %}

Since the string "foobar" is not quoted in this example, Twig will not
find it as the value and will throw an exception. We should try to catch
it and return a somewhat helpful message instead.